feat(sim): protocol v27 — fuel burns as itself, damage speaks per side
W0.7 of the release master-program: the single pre-network protocol
bump, batching every wire change before the transport lands.

- `TankSnapshot.fuel_fire` (v27): a fuel-tank component holed by the
  internal path lights a FUEL fire, distinct from the engine's own —
  a tank can blaze with a healthy engine. Presentation and the audio
  crackle bed key on either flame.
- The damage-event component mask widens u16 -> u32: ids 17+ were
  silently maskless, which capped the interior at sixteen components
  and kept the loose rounds visual-only. Every id now owns a bit
  (collision-locked test).
- Suspension honesty (behavioral, no wire): a penetration into the
  suspension module degrades the STRUCK side's band only — the
  torsion bars are independent left/right assemblies and a shell in
  the left bank no longer cripples the right. HE still takes both.
  (The planned slot split proved a 68-site refactor for the same
  gameplay truth; the hit-side degrade achieves it on the wire we
  already have. The radio consumer moves to N2, where per-viewer
  filtering can express it — as the plan's N2 package already states.)

v27 fixtures regenerated per the testing-and-regression procedure;
v26 fixtures and old-version rejection tests stay. The populated
combat fixture carries fuel_fire=true so the flag is byte-locked.

+++ crates/foundation/game_core/src/damage.rs
@@ -115,5 +115,5 @@ pub struct DamageEvent {
     /// (protocol v26). Bit `n` represents `DamageComponentId(n + 1)`; legacy ModuleSlot masks
     /// remain the HUD/repair compatibility layer.
     #[serde(default)]
-    pub damaged_components_mask: u16,
+    pub damaged_components_mask: u32,
 }
